+++ __tests__/workflow/step.test.ts
@@ -68,4 +68,33 @@ describe("step", () => {
             }
         });
     });
+
+    test("make non managed", () => {
+        expect(
+            new Step({
+                run: "# github-actions-managed: true\nmyline"
+            }).makeNonManaged().data
+        ).toEqual({
+            run: "myline"
+        });
+
+        expect(
+            new Step({
+                with: {
+                    script: "// github-actions-managed: true\nmyline\n  other",
+                    "github-actions-managed": true
+                }
+            }).makeNonManaged().data
+        ).toEqual({
+            with: { script: "myline\n  other" }
+        });
+
+        expect(
+            new Step({
+                run: "myline\nother"
+            }).makeNonManaged().data
+        ).toEqual({
+            run: "myline\nother"
+        });
+    });
 });