+++ src/workflow/step.ts
@@ -49,6 +49,37 @@ export class Step {
         return false;
     }
 
+    makeNonManaged(): Step {
+        if (!this.isManaged()) return this;
+        if (
+            this.data.run &&
+            this.data.run.includes("# github-actions-managed: true")
+        ) {
+            const lines: Array<string> = this.data.run
+                .split(/\r?\n/)
+                .filter((l, i) => i || l.trim())
+                .filter(l => l.trim() !== "# github-actions-managed: true");
+            this.data.run = lines.join("\n");
+        }
+        if (
+            this.data.with?.script &&
+            this.data.with.script.includes("// github-actions-managed: true")
+        ) {
+            const lines: Array<string> = this.data.with.script
+                .split(/\r?\n/)
+                .filter((l, i) => i || l.trim())
+                .filter(l => l.trim() !== "// github-actions-managed: true");
+            this.data.with.script = lines.join("\n");
+        }
+        if (this.data.with?.["github-actions-managed"]) {
+            delete this.data.with["github-actions-managed"];
+            if (Object.keys(this.data.with).length === 0) {
+                delete this.data.with;
+            }
+        }
+        return this;
+    }
+
     makeManaged(): Step {
         if (this.isManaged()) return this;
         if (this.data.run) {